+++ app/src/androidTest/java/com/example/buttonapp/autotesting/inagraph/INAGraphBuilder.java
@@ -1,16 +1,16 @@
 package com.example.buttonapp.autotesting.inagraph;
 
-import androidx.test.uiautomator.UiDevice;
-import androidx.test.uiautomator.UiObject;
-import androidx.test.uiautomator.UiObjectNotFoundException;
-
-import com.example.buttonapp.autotesting.inagraph.actions.Action;
-import com.example.buttonapp.autotesting.inagraph.actions.ActionFactory;
-
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Random;
+
+import androidx.test.uiautomator.UiDevice;
+import androidx.test.uiautomator.UiObject;
+import androidx.test.uiautomator.UiObjectNotFoundException;
+import com.example.buttonapp.autotesting.inagraph.actions.Action;
+import com.example.buttonapp.autotesting.inagraph.actions.ActionFactory;
 
 public class INAGraphBuilder {
 
@@ -56,7 +56,7 @@ public class INAGraphBuilder {
     }
 
     public void createActions(Node node, UiDevice device) {
-        Map<UiObject, Action> actions = ActionFactory.createActions(device,  new Random().nextInt());
+        Map<UiObject, Action> actions = ActionFactory.createActions(device,  new Random().nextLong());
         node.getControls().addAll(actions.keySet());
         node.getAvailableActions().addAll(actions.values());
     }
@@ -110,7 +110,7 @@ public class INAGraphBuilder {
 
     public boolean isSameNode(Node currentNode, UiDevice device) throws UiObjectNotFoundException {
         boolean result = true;
-        List<Action> actions = new ArrayList<>(ActionFactory.createActions(device, new Random().nextInt()).values());
+        List<Action> actions = new ArrayList<>(ActionFactory.createActions(device, new Random().nextLong()).values());
         for (int i = 0; i < actions.size() && result; i++)
             result = (result && currentNode.getAvailableActions().contains(actions.get(i)));
         result = (result && currentNode.getAvailableActions().size() == (actions.size()));
